+++ frameworks/Rust/axum/Cargo.toml
@@ -1,8 +1,20 @@
 [package]
-name = "axum_techempower"
+name = "axum"
 version = "0.1.0"
 authors = ["Dragos Varovici <dvarovici.work@gmail.com>"]
 edition = "2018"
+
+[[bin]]
+name = "axum"
+path = "src/main.rs"
+
+[[bin]]
+name = "axum-sqlx"
+path = "src/main_sqlx.rs"
+
+[[bin]]
+name = "axum-bb8"
+path = "src/main_bb8.rs"
 
 [dependencies]
 rand = { version = "^0.8", features = ["small_rng"]}
@@ -12,15 +24,19 @@ async-trait = { version = "0.1" }
 futures = { version = "^0.3" }
 futures-util = { version = "^0.3" }
 dotenv = { version = "^0.15" }
-
 serde = { version = "^1", features = ["derive"] }
 serde_json = { version = "^1" }
 serde_derive = { version = "^1" }
-
 axum = { version = "^0.3" }
 tokio = { version = "1.0", features = ["full"] }
 hyper = "0.14"
 tower = { version = "0.4", features = ["util"] }
 tower-http = { version = "0.1", features = ["set-header"] }
-
 sqlx = { version = "^0.5", features = [ "postgres", "macros", "runtime-tokio-native-tls" ] }
+bb8 = "0.7"
+bb8-postgres = "0.7"
+tokio-postgres = "0.7"
+
+[profile.release]
+lto = true
+codegen-units = 1
