+++ frameworks/Rust/axum/axum-bb8.dockerfile
@@ -1,5 +1,7 @@
 FROM rust:1.55-slim-buster
 
+ENV AXUM_TECHEMPOWER_DATABASE_URL=postgres://benchmarkdbuser:benchmarkdbpass@tfb-database/hello_world
+
 RUN apt-get update && apt-get install -y --no-install-recommends \
     libpq-dev pkg-config libssl-dev \
     && rm -rf /var/lib/apt/lists/*
@@ -15,4 +17,4 @@ RUN cargo build --release
 
 EXPOSE 8000
 
-CMD ["./target/release/axum"]
+CMD ["./target/release/axum-bb8"]
