+++ frameworks/Rust/axum/rustfmt.toml
@@ -1,0 +1,2 @@
+max_width = 89
+reorder_imports = true
