--- frameworks/Rust/rocket/.env
@@ -1,1 +1,0 @@
-ROCKET_BENCHMARK_DATABASE_URL=postgres://benchmarkdbuser:benchmarkdbpass@localhost/hello_world
